common/string_utils: use handwritten stringFormat
Using spdlog::fmt (which is really just fmtlib) is overkill. Our
codebase only uses `{}`, and so this commit implements exactly that
(along with the ability to escape values). Experimentally this gives us
a 15% compilation speedup.

It also moves string formatting to its own file, because of how often it
is included. Finally, this change is conducive to adding clang-analyzer
to clang-tidy, as spdlog has an unitialized object warning.