storage/buffer_manager: use locking queue
The lock-free queue is the source of an ASAN failure. Specifically, the
queue does not guarantee strict FIFO circumstances in the name of speed.
This means that our code that loops through the queue is incorrect.

The goal is to replace this queue-based design soon regardless, and
hence we don't want to spend more time on this.

Also, this commit removes a useless logger in the buffer manager.