mac: fix explicit symbol exports
Exception information necessary for properly catching exceptions, seems
to not be exported on Mac unless we export the exception, which seems to
be different from Linux & Windows.

In additional, Mac issues a warning when building the Python module
(which itself builds as a shared library). The issue is that the python
module links with our static library, and the static library does not
export symbols (since it doesn't do anything for static libraries).

However, this mismatch in visibility causes a warning. Technically, we
shouldn't build a shared library with a static library since the static
library isn't position-independent code, but we don't really have other
options unless we embed all of the object files?

We should probably link the python library against `libkuzu.so` and ship
it alongside the python bindings. However, this will be delayed until
after the release.

Ref #2136.