cmake: enable LTO
Link-Time Optimization is a technique whereby the linker performs
optimizations at the final step, right before generating the final
executable or shared library. This provides significant performance and
binary size optimizations, since the linker has access to all the
context needed, unlike the individual optimizer passes, but in exchange
linking takes a lot of time and memory.

Based on some reading and some experimentation, Unity Builds, though
they can offer similar benefits, are not nearly as good for
optimization, and instead are primarily used for speeding up compile
times. Additionally, Unity builds are harder to get right and harder to
debug. LTO on the other hand is a very simple change, and seems to be
the main area of focus right now.

In order to make this work on Windows, we needed a small fix to how
ANTLR4 is built.

Closes #2036.