cmake: rewrite grammar generation
The script to generate our grammar files is now written in CMake's
scripting language, and as such is cross-platform. Additionally, the
script is run as part of the build. We commit a copy of the Cypher.g4
file so that Java is not a build dependency of Kuzu: specifically, if
the grammar has not changed, the generation command will do nothing.

Additionally, a sanity check that the generated files are up to date has
been added to CI.