reader/csv: avoid lseek for getting offset
Currently we use lseek() after every row in the parallel CSV reader to
get the file offset. I thought this would be very cheap, and figured it
was easier than tracking it ourselves. However, it is not cheap at all,
as demonstrated by a flamegraph, and it is not hard to track ourselves
either.