common: add MPSC Queue
This queue is core to the new parallel hash index. It implements the
design in this [blog][1], which is also used by Rust. The design is quite
simple, and producers are wait-free. The main disadvantage is it is
non-blocking, but adding blocking pop() support is tricky, and currently
not needed.

  [1]: https://www.1024cores.net/home/lock-free-algorithms/queues/non-intrusive-mpsc-node-based-queue