Fix reading strings from the hash index overflow file
When there is multi-threaded contention, data was being appended to the
result multiple times because of how optimisticRead behaves.
Replacing the relevant portion of the string instead of appending makes
sure that the result is correct and overwrites the null data from
previous runs when the optimisticRead finally manages a successful read.

Fixes #2916