Fix cleaning up the test directory on windows/macos CI
The windows and macos CI run multiple tests in different build directories as the same user.
Since #3290 this means that the tests use the same temporary directory, and the cleanup_test may remove tests from other runs.
This uses the RUNNER_TEMP environment variable to create job-specific test directories on the runner machines