Fix data race in eviction queue
The candidate may be evicted by another thread between the first time it
is checked and when it is loaded, in which case the pageState lookup
will cause UB as the fileIdx will be out of bounds.

Also changed pagesEvicted to evictablePages for clarity, since they don't get
evicted until the end of the function and the counter just tracks the number
of candidates found.