Set the SelectionVector to STATIC_FILTERED when scanning cached rels if possible
bitset access is relatively slow, and may not be necessary
There are also some optimizations possible to take advantage of not having to scan the SelectionVector's buffer (though I'm not sure if we're actually taking advantage of it right now).
And I avoided the unnecessary creation/destruction of a std::shared_ptr in RelTableScanState::setNodeIDVectorToFlat by using getSelVectorUnsafe instead of getSelVectorShared.