Fix stop threshold for BM evictions (#5176)
* Fix stop threshold for BM evictions

Sets the stop threshold for eviction to be exactly twice through the eviction queue

Previously we were not only skipping counting empty slots, but were not taking into account other threads evicting at the same time. We should fail after going through the eviction queue twice regardless of the number of other threads which are also evicting.
This significantly improves failure time when a large number of threads are trying to evict at the same time

* Sleep briefly and retry prior to failing to allocate BM memory