fsync() the WAL on open
It's generally wise to ensure that whatever set of changes are being
recovered from the WAL are actually durable before they're processed. By
a strict interpretation of the UNIX spec, the fsync() on open doesn't
_have_ to actually do anything, but in practice on Linux and the common
ext4/xfs it does, and thus avoids weird crash->recover->power
loss, and now storage is now ahead of the WAL sorts of weirdness.